Added frontend components
