+++ frontend/civic-voice/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "municipality-complaints",
+  "name": "civic-voice",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -9,16 +9,16 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.126",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "axios": "^1.8.4",
-    "bootstrap": "^5.3.4",
+    "bootstrap": "^5.3.5",
     "jwt-decode": "^4.0.0",
     "react": "^19.1.0",
     "react-bootstrap": "^2.10.9",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^7.4.1",
+    "react-router-dom": "^7.5.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
