+++ frontend/civic-voice/src/App.tsx
@@ -1,4 +1,5 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+
 import Home from "./pages/Home";
 import Login from "./components/login";
 import Register from "./components/Register";
