+++ frontend/civic-voice/src/components/protectedRoutes.tsx
