+++ frontend/civic-voice/src/index.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import reportWebVitals from "./reportWebVitals"; // ✅ This will pick reportWebVitals.ts
+import 'bootstrap/dist/css/bootstrap.min.css';
+import App from "./App";
+
+
+const root = ReactDOM.createRoot(
+  document.getElementById("root") as HTMLElement
+);
+
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
+
+reportWebVitals();
